Bring down local variables that came before local functions.
Summary:
Problem: Not able to understand this function without a lot of effort.

Let's reorganize it as local functions before local variables, so that local variables aren't implicitly captured all over the place. (This is the coding style we agreed on.)

This will allow us to factor out these functions in the future.

I'm not done yet. There are five more local variables used throughout the nested functions.

Reviewed By: kinto0

Differential Revision: D41778054

fbshipit-source-id: b97654eb858b6f5ea2851bc0b35f51397248424d