Make `attribute_from_class_name` return readonly attribute for readonly class.
Summary:
Problem: The previous diffs passed in a flag indicating whether an attribute was accessed in a readonly type.

Use that type.

This changes the behavior for direct calls `GlobalResolution.attribute_from_class_name`. Will change other APIs (and the type inferred for attribute lookups `foo.bar`) in coming diffs.

Reviewed By: sinancepel

Differential Revision: D41778055

fbshipit-source-id: 03175bc853aa103966d287f3b6ae14aee60c6305