Derive `show` for key types in `AttributeResolution`.
Summary:
Problem: Until we get a debugger, we have to resort to log-dumping. But we have many types that don't derive `show`, meaning that we have to add and remove them locally each time.

Solution: Derive `show` and check it in.

The only real counter-argument to not deriving and exposing `show` is that it breaks our data-hiding, since we aren't exposing the implementation. However, given the lack of a debugger and the severe difficulty of debugging our OCaml code (easily several hours), I don't think that's worth hampering everyday debugging.

Can discuss in the Engineering sync about a general coding style.

Reviewed By: sinancepel

Differential Revision: D41816513

fbshipit-source-id: 7600b5c6e8e379c2e6fe933ec130c00c94172689