Emit error when assigning to a readonly attribute.
Summary:
Problem: This stack switches over to emitting readonly errors in the type checking analysis. We haven't yet emitted errors when assigning to readonly attributes.

Note that we check for `is_class_toplevel` because we don't want to error about the attribute initialization in the class.

Note: I've added a couple of test cases that don't work as expected - these are cases where we mix `ReadOnly` and `Type`. `forward_assignment` uses janky, hand-rolled code for checking if something is being accessed through `Type[...]`. The fix will be to make that use `Type.class_data_from_type`. Not prioritizing that right now.

Reviewed By: sinancepel

Differential Revision: D41778058

fbshipit-source-id: 31cfd072b348d783b4342d692aa191400548c8c4