Stop filtering error when calling a mutating method on a readonly object.
Summary:
Problem: Previous diffs made sure that a method looked up on a readonly object had the `self` argument as readonly. That in fact did lead to an error when the method had a mutable `self` parameter. However, we used to filter all `self` mismatch errors blindly.

Stop doing that for readonly objects.

Note that I'll be switching over to the specialized readonly error messages at the end of the stack. Right now, we just emit the generic incompatible-parameter error.

Reviewed By: sinancepel

Differential Revision: D41816511

fbshipit-source-id: 5ded200ebee8402531314ad8affc834bbbaa1525