Allow binding type variables to readonly types.
Summary:
Problem: We were saying that beyond the obvious cases, `ReadOnly _ </: _`. This meant that `ReadOnly[int] <: T` led to a type mismatch and `T = Any`.

Move the readonly "impossible" case after the type variable cases.

Add unit tests to catch the branches that come in between, such as `Any`, `TypeVar`s on the left or the right side, etc.

There are probably a couple of other cases where we've hardcoded `object` as top, but I think it's fine to wait for actual user requests to fix them.

Note: The error message incorrectly says `"1st positional only parameter"` instead of `"... argument"`. I fixed that for the old readonly-specific errors. Need to fix it for the generic errors. Will do it later in the stack.

Reviewed By: stroxler

Differential Revision: D41847827

fbshipit-source-id: 446413190282201adcfc44cdfefa18f25a265888