Refactor the model query generation
Summary:
In this diff, we split the `apply_all_queries` function into smaller functions easier to reason about.

This function was doing many things:
* Partition queries into callable queries, attribute queries, global queries
* Fetch callables/attributes/globals
* Perform a map-reduce with multiprocessing on callables/attributes/globals
* Iterate on queries and callables/attributes/globals to build a `ModelQueryRegistryMap`

All this was implemented with a few generic functions that were taking other functions as parameters, it was really messy and hard to read.

In this diff, we implement a generic module that can generate models for a given type of query. Then, we implement the different logic for callables, attributes and globals. Finally, the new `apply_all_queries` simply threads everything together.

I have also improved names of function drastically. `apply_queries` was very generic.

Reviewed By: litho17

Differential Revision: D41827003

fbshipit-source-id: 76ae28ef5a7d2d30d35ac9e9b47c71d34fe972dd