Rename connect_and_subscribe
Summary:
Nothing about `launch_and_subscribe_handler` is actually
public except its `background.Task` interface.

That said, we don't usually make rigorous use of public/private
naming conventions in Pyre, which I think is fine - the leading
underscores can hurt readability and outside of libraries there's
not that much benefit.

Instead, we tend to mark "high-level" logic without leading
underscores, and "low-level helpers" with them. But that makes it
confusing to have `_try_connect_and_subscribe`, which is higher-level
than `subscribe`, marked with an underscore.

Also, the `try` isn't really descriptive here - I only like using
that in a name if there's a non-try version of the same function
(i.e. when we split the exception-handling logic into a separate
layer) which is not happening here.

Reviewed By: grievejia

Differential Revision: D41878472

fbshipit-source-id: c49ce80789ae7956a68470a46445d201137097e4