Better API for `FetchCallables.t`
Summary:
The `FetchCallables` module is responsible for finding all callables in the current project and categorizing them (internals, stubs, etc.).
The current API is a bit confusing because it has a `get_callables` which returns all callables **except stubs**. This can be the source of errors.
To prevent this, this diff changes the API to make it a lot clearer when we return stubs or not.

Reviewed By: alexkassil

Differential Revision: D41864022

fbshipit-source-id: f2256ce0b0254a17cb8be525938120df51c952cf