Handle not-isinstance for readonly types.
Summary:
Problem: We previously handled the case where we were refining `x: ReadOnly[Child] | int` for `isinstance(x, Child)`. However, we didn't refine `x` in the else branch to be `int`.

That requires us to handle `assert(not isinstance(...))`.

Change the `partition` function to ignore readonlyness when deciding which types go in the if vs else branch.

Note that this wouldn't have been sufficient to refine `x: ReadOnly[Base] | int` for `isinstance(x, Child)`. See D41893176 for what was needed.

Reviewed By: stroxler

Differential Revision: D41895131

fbshipit-source-id: 158f178ab7b2911b2c85d5651a569e617707a1dd