Split the parsing and translation into an AST
Summary:
The parsing logic in `generator.mly` is currently mixed up with the translation into a proper AST.
In this diff, we split those up, by introducing a `ParserStatement` module which defines statements returned by the parser.
Then, in `parser.ml` we translate parsed statement into the proper AST.

For now, this is a no-op. This will allow to properly expand format strings without requiring post processing.

Reviewed By: grievejia

Differential Revision: D41987145

fbshipit-source-id: 2e61d38ef8853ec1b7e0e7e307b61ea2b193562a