Migrate error variant for incompatible variable type.
Summary:
Problem: We are emitting the regular "Incompatible variable type" error. Want to specialize the error messages for readonly, so that we have a special error code for them. That way, we can easily track or suppress readonly violations.

We were specializing `incompatible_type` record for readonlyness. Switch to the type-based mismatch type.

Next diff: Use the variant when emitting errors.

Reviewed By: connernilsen

Differential Revision: D41987891

fbshipit-source-id: 43838b6ba73be0e2d39f2c8821b57df2019b5ce3