When creating triggered partial sinks, also remember the matched sources
Summary:
Such that later when these triggered partial sinks are matched, we know all
sources (that have been matched), in order to create first hops to these
sources. Then, we can show subtraces in the Zoncolan UI to show all source
traces.

Context: Given a multi-source rule from SourceA and SourceB to SinkC, we create
a triggered sink SinkCA if finding a flow from SourceB to SinkC, such that later we
only need to match SourceA with SinkC.

Reviewed By: arthaud

Differential Revision: D41881616

fbshipit-source-id: cf01f1c986d85dd68b74a848fede337146cd8aba