Use {Reference,Identifier}.Map.Tree in module.ml code
Summary:
The data we put in shared memory is always of type `Key.Map.Tree.t`
rather than `Key.Map.t`, because the former stores a comparator
statically in a functor whereas the latter stores it as a closure and
is therefore not serializable.

So, it would be simpler if we always operate on the `Key.Map.Tree.t`
data; the operations are the same but we need to access them statically
rather than through `Map.function_name` with indirection.

This is important to me now because I want to upgrade to Core 0.15.x (which
will block any experimentation with Ocaml 5.0), and this release of Core
removes the `to_tree` conversion that we've been relying on. I'm sure it
would still be possible to covert, but I think this fix is cleaner anyway.

Reviewed By: jasontatton-meta

Differential Revision: D41973214

fbshipit-source-id: 4cd23feec181501620b2a142bad2aabae8d2806a