Don't alias Errors in configurationTest.ml
Summary:
Similar to the previous diff: for reasons I'm not entirely clear on, ocaml
currently allows us to bind two Error names (the TaintConfiguration.Error
module and the Result.Error variant) in the same scope, even though this
is pretty confusing.

I'm not clear why, but bumping Core to 0.15.x causes this to stop working;
it is a bad idea in the first place so this is fine by me.

Fix is to use TaintConfiguration.Error explicitly so that we can continue
to `open Result` in the tests.

Reviewed By: arthaud

Differential Revision: D42005654

fbshipit-source-id: 6fbeba46a712c4a947fa2735bd3af6eda2b02362