Do not open Result in taintConfiguration.ml
Summary:
We used `open Result` to get access to `Result` infix operators in this
module. This leads to some pretty confusing name collisions, because we
have both `TaintConfiguration.Error` and `Result.Error` in scope at the
same time.

I'm actually not sure why this works - I guess type inference is good enough
to distinguish them - but after we bump the Core version to 0.15.x this breaks
and we become unable to access `TaintConfiguration.Error`.

So, to fix this I'm just binding the infix operators directly at the top-level
(in the process accessing `Option.(>>|)`, which is used exactly once, directly
rather than via an `open Pyre`) and then avoiding `open Result`.

Reviewed By: arthaud

Differential Revision: D42005658

fbshipit-source-id: 015b2d64def1c3f2c92ad45c24eacd10527155d3