Flip Pysa code to use Target.Map.Tree.t
Summary:
The reasoning for this change is the same as for D41973214: in
Core 0.15.x there's no to_tree function in `Target.Map` to convert
between `Target.Map.t` and `Target.Map.Tree.t`; since we need the latter
for shared memory, we should use it everywhere (there's also a tiny
memory savings, since the tree version gets the comparator statically rather
than via a closure).

Unfortunatley, in 0.14.x there's no `Target.Map.Tree.merge_skewed`; as a result
we have to hack one by converting back and forth; in 0.15.x this problem goes
away and I'll land everything at once so this won't actually be a problem
on trunk.

Reviewed By: grievejia

Differential Revision: D42005657

fbshipit-source-id: d6a3188a2f7502cfba091495dcb8b0ef804f5219