Parse `read_from_cache` and `WriteToCache` in model queries
Summary:
# Context
Model queries can be quite slow to generate models, especially if the number of queries is high (>10k).
In the following diffs, we will implement a new syntax for model queries that will make them faster.

# Solution
We introduce a new type of model queries, using the `WriteToCache` directive.
This can be used to build a mapping from arbitrary string to a set of targets (functions, methods, attributes or globals).
Then, a new constraint called `read_from_cache` can be used in regular model queries to find targets for a given string.

This can transform `N` model queries with slow constraints into a single expensive query and `N` very cheap queries.

# Example
Imagine having the following queries:
```
ModelQuery(
  ...
  where = [
    AnyOf(cls.extends("my_module.Foo"), cls.extends("other_module.Bar")),
    fully_qualified_name.matches("\.MyClassOne\.method$"),
  ],
  model = ...
)
ModelQuery(
  ...
  where = [
    AnyOf(cls.extends("my_module.Foo"), cls.extends("other_module.Bar")),
    fully_qualified_name.matches("\.MyClassTwo\.method$"),
  ],
  model = ...
)
ModelQuery(
  ...
  where = [
    AnyOf(cls.extends("my_module.Foo"), cls.extends("other_module.Bar")),
    fully_qualified_name.matches("\.MyClassThree\.method$"),
  ],
  model = ...
)
```

Each model query is expensive to compute since it uses `cls.extends`, which needs to check the class hierarchy.
To make this faster, we could use the following queries:
```
ModelQuery(
  ...
  where = [AnyOf(cls.extends("my_module.Foo"), cls.extends("other_module.Bar"))],
  model = WriteToCache(kind="example", name=f"{class_name}:{function_name}")
)
ModelQuery(
  where = read_from_cache(kind="example", name="MyClass:method",
  model = ...
)
ModelQuery(
  where = read_from_cache(kind="example", name="MyClassTwo:method",
  model = ...
)
ModelQuery(
  where = read_from_cache(kind="example", name="MyClassThree:method",
  model = ...
)
```

This will be a lot faster to generate.

# This diff
In this diff, we implement the parsing of `WriteToCache` and `read_from_cache`.

Reviewed By: alexkassil

Differential Revision: D41988092

fbshipit-source-id: 095b1bb647a3bb2f31fffd48e1ceeabc0cda4042