Add definition for undefined function.
Summary:
Problem: The test code referred to `await_list`, but there was no such function.

Technically, this doesn't matter for the analysis, since it just looks at whether awaitables are flowing into some function argument, not whether the function is defined. (That's the responsibility of the regular type checking analysis.)

Still, to avoid future head-scratching, it's better to have the function defined.

Reviewed By: kinto0

Differential Revision: D42050003

fbshipit-source-id: acb814630d5f64f94329c5ce27c7ae7f1a27b539