Split constraint checking and annotation generation in different functions
Summary:
To implement the `WriteToCache`/`read_from_cache` model queries, we will need slightly different logic than before.
Instead of generating taint annotations, we will need to generate a cache of targets that matched the constraint.
To prepare for this, we need to split further the implementation of the model query generation.

Reviewed By: alexkassil

Differential Revision: D42058506

fbshipit-source-id: 83852ca5ab2a3d76a23e87e3399fb44ecea83876