factor out server start commands from persistent_test
Summary:
the persistent server is not the only one needing this boilerplate. to be reusable in the next diff, I factored out the common testing commands to its own file. It can't be added to setup.py or there would be a circular dependency so I called it server_setup.py - open to name suggestions

it was originally supposed to be for D42058661 however we don't really have enough integration test scaffolding for the time being. I still think it's worth moving this code to a common place in case a test similar but simpler than D42058661 comes around

Reviewed By: stroxler

Differential Revision: D42058663

fbshipit-source-id: 0b67f5f630fe63c511a5e3a1e276399879c29c53