Send client state to daemon on restart
Summary:
If the pyre client receives any user input while the daemon is down, they fail. This is ok for most requests since they can just be retried. However, document open requests are not requests triggered by the user intentionally.

If the server is down when the open request comes in, unsaved changes will not work for that file. This diff makes a change that resends these on server cold boots.

Alternative options:
- Instead of this callback approach, we could put the logic in launch_and_subscribe_handler. I chose not to because I thought this should be the responsibility of the language server, not the pyre daemon launch/subscribe handler. it couples this behavior to the handler and means other clients can't control it independently. For example: I only added this callback in code_navigation because I found the persistent server was able to receive didopens while building just fine. This is because there's a difference in the overlay creation semantics between codenav/persistent.

Reviewed By: grievejia

Differential Revision: D42058662

fbshipit-source-id: 31d5e9116f32a496cd83a7680b99b68595f054ee