Format responses better
Summary:
Context:
hover/gotodef is able to return multiple responses if there are multiple conflicting modules. Our [current way of formatting hover](https://fburl.com/code/1op0nczw) doesn't make a ton of sense where we show all the types then all the docstrings.

This diff changes hover to show each type/docstring together. This fixes a bug with docstrings in which "None" is printed instead of nothing:
https://pxl.cl/2nbcq

To do that, we we are now returning a `HoverResponse` (list of `lsp.PyreHoverResponse`) from `async_handle_hover_request` instead of a single `lsp.PyreHoverResponse`. We can get rid of `HoverContent` since it's a duplicate of `PyreHoverResponse`.

Reviewed By: sinancepel

Differential Revision: D42105053

fbshipit-source-id: fea44edc76ea8baa592d80ea4adc1f72da9a4013