Replace `PysaReference` with `SerializableStringMap`
Summary:
We already define a `SerializableMap` in `data_structures`. Instead of implementing `PysaReference` with `Core.Map.Tree`, we could just instantiate `SerializableMap` for `string`.

The downside is that the `SerializableMap` is missing a few functions, but these can be implemented easily.
Another downside is that the API for `SerializableMap` is based on `Caml.Map`, which is not as nice as `Core.Map` since it does not use named parameters and the order of arguments is sometimes inconsistent.
The upside though is to have a generic data structure implementation which we can easily replace in the future.

Reviewed By: stroxler

Differential Revision: D42088048

fbshipit-source-id: 1d082a73127da2d652de69ae6953891502ba5de0