Do not consider classes implementing `__await__` as awaitables.
Summary:
Problem: Our aim is to prevent production errors when users forget to `await` awaitable expressions. To make this problem concrete and actionable, we set the goal of soundly preventing `RuntimeWarning: coroutine xx was never awaited`.

This covers all of the user reports we have received.

However, we also consider an object to be awaitable if its class satisfies the `Awaitable` protocol (has `__await__` of the right type).

Reviewed By: sinancepel

Differential Revision: D42154550

fbshipit-source-id: 3d25d83b4d48a46509d1c572ff17ab45ec69b4dc