end-to-end implementation of query API on Python for superclasses
Summary:
This diff serves as a prototype of how we can build a more programmable query API for the CodeNav server. The initial use case is to power the `Superclasses()` query in an ergonomic manner for the LibCST integration we have planned, but focus is on design.

Want to get feedback here on the overall approach and make sure we agree on the ergonomics :)

Reviewed By: grievejia

Differential Revision: D42078467

fbshipit-source-id: 2dfe974a6a751cdcf2fff48995a88422c5362f71