maintain a set of open files
Summary:
Maintaining a set of open files will be necessary to support Pyre API calls for linters. This diff adds the APIs to communicate open/closed files to the server.

As of the time of writing, these files being open/closed shouldn't affect server behavior, but this gets us ready for the bright new future.

In addition to sending requests, we keep a set of files - in the future, if we want to handle server failures more gracefully, we'll want to re-transmit all open files to a newly created server.

Reviewed By: grievejia

Differential Revision: D42172462

fbshipit-source-id: a817ec45dac03366341c9ee09f9f9dd1304e6477