start supporting opening and closing files without an overlay
Summary:
Context: For the API usecase, we'll want to mark files as open to trigger Buck runs.
However, there's no associated overlay - we'd want the root environment to be used for any type analysis. The initial approach for open files required an overlay as we assumed an associated IDE would exist, but this breaks down quickly for the API use case, so let's backtrack :)

Reviewed By: grievejia

Differential Revision: D42170512

fbshipit-source-id: 1ed0f98627aa8e6bb5bd3014cb2da4767f5f0ac1