Don't emit error when assigning to readonly attribute of a mutable object.
Summary:
Problem: We don't want users to assign to an attribute of a readonly object (`my_readonly_object.attribute = 42`).

But it should be ok to assign to an attribute that is explicitly marked as `ReadOnly[str]`, since the only thing forbidden there is that you can't mutate the value it points to.

+ Also add a test for the constructor when someone assigns `self.readonly_attribute = initial_value`.

Reviewed By: connernilsen

Differential Revision: D42182131

fbshipit-source-id: 52a08a9608a3f767fbaaa3e1d0b3d6bc6e7e607a