print codenav server logs on `pyre info`
Summary:
only the typechecking server logs are printed on `pyre info` (even if there is no typechecking server up). This change makes it so the codenav server logs are also shown (even if there is no codenav server up).

I would like to automate automate this more in the future to only display the necessary logging files if we want. However, I think that would be pretty difficult on the command line. Some ideas and why they don't work too well:
- `pyre info --flavor=code_navigation`: not that ergonomic
- `pyre info` searching upwards for the necessary configurations: we currently don't even search up for `.pyre_configuration.codenav`s so I'm not sure we want this logic just for this

Reviewed By: sinancepel

Differential Revision: D42117597

fbshipit-source-id: d17db4d395536460389f251e91cf461fb6b053eb