Refactor incremental_style -> track_dependencies
Summary:
The configuration field named `incremental_style`'s name was
misleading - it was an artifact from an older architecture that
had two styles of dependency tracking, but we haven't used shallow
tracking in years, and about 6 months ago we actually removed the
ability to use shallow tracking entirely.

Instead, what that field controls today is whether we try to track
dependencies at all; any one-shot command (check, analyze, infer, etc)
can be run without dependency tracking, so we use this flag to skip
some overhead.

Let's just make it a boolean named `track_dependencies` to make this
clear. Context on why I was looking at this is that we are still doing
some dependency tracking work even when tracking is turned off due to
the way code is factored, and I'm considering trying to cut that down
as a perf optimization.

Reviewed By: grievejia

Differential Revision: D42369578

fbshipit-source-id: a1f214ad3b1aec1abd9f8600a3bab40a52f9755b