Bump GC space overhead a bit
Summary:
This was recommended in an article on tuning the GC for data processing jobs:
https://md.ekstrandom.net/blog/2010/06/ocaml-memory-tuning

I tried all the other tunings recommended, this is the only one that made much difference, it seems
to give us a couple percent boost - not necessarily stat sig in any one servicelab test, but it's been
reliable on rebase so I'm pretty certain the effect is real.

More overhead makes sense when you care a bit more about wall time than RAM efficiency. I use
the lower end of the range for the long-lived main process and the higher end for workers, which are
ephemeral (in Pyre this doesn't matter so much, but in Pysa where we perform many map reduce
iterations the fact that workers are ephemeral could be important - it means the RAM use of workers
shouldn't really grow over time so we can afford quite a bit of overhead).

Reviewed By: sinancepel

Differential Revision: D42350782

fbshipit-source-id: 530a6a2350a66c0c8129af79b341054a67c711bf