Remove Gc.stat from the non-debug path
Summary:
When performance testing Pyre against Ocaml 5.0.0, we hit a
performance regression. Some of this regression involved typeCheck,
but after looking at the times for a while I realized a great deal
of it was actually not coming from any of our processing at all!

It turns out that on Ocaml 5.0.0, at least in our environment, the
Gc.stat call can take a long time - on my devserver, type checking
one of our big repos (where type check took ~450s), the Gc.stat
call took almost 50s(!!) to complete. This is compared to essentially
zero time - at most maybe 1s - in 4.14.

This is a regression we might want to report to ocaml in case it
isn't previously known; in the meantime though we hardly ever use
these numbers (maybe never?) so let's just put them behind a debug
flag, along with the `Memory.report_statistics` calls - both of these
bits of telemetry can be handy in very specific situations but don't
need to be part of the ordinary production path.

Reviewed By: sinancepel

Differential Revision: D42387480

fbshipit-source-id: a8bb08dce0c7c85d7845b767aa1b56f0f1070c15