Double chunk size to 10K in type check
Summary:
I was thinking to use chunks per worker rather than chunk size,
because in theory chunks per worker is better at giving us dynamic
control of sharding for projects of variable size.

But when I pushed chunks per worker to 1, the process crashed. I
suspect this was due to us trying to serialze too large an object,
likely in the collected map reduce, and this makes me realize that using
chunks per worker in type check might be risky if it can cause the
dependency tracking aggregation to overflow.

The problem here is that if we fix chunks per worker, natural repository
growth could still eventually trigger Pyre to start crashing, which is not
a great state to be in.

Instead, we might be better off just cranking up the chunk size, but
still keeping it fixed. I believe that typeCheck is the *only* case where
this is an issue, since it does so much more work; collecting defines and
postprocessing both seem to be safe to do using 1 chunk per worker.

Note: I tried 20K and 30K; both of them help but 10K works best (about
5% faster) - likely this is because type check times are skew, so we do want
at least a few chunks per worker unlike in definition collection and postprocessing.

Reviewed By: sinancepel

Differential Revision: D42376812

fbshipit-source-id: 2ba8406d4b2a08f80dc5b58cf6c8c14fb5190bc6