Log no error as null instead of 'None'
Summary:
For hover and go-to def, we stringify the error before logging, which turns
errors of None into the string 'None'. This is different from what happens when
we don't set any errors (like in didSave) where logs show up as `null`. This
turns hover and go-to def into `null` as well.

Reviewed By: jasontatton-meta

Differential Revision: D42391071

fbshipit-source-id: ea8f63725ecf23ecb2cbf3e05645e54f6b4a7bde