Let PartialSinkConverter be a map from partial sinks to sources and (other) partial sinks
Summary:
As opposed to *a list of* sources and (other) partial sinks. For example, consider rule
```
{
  SourceA, SourceB
  SinkC
}
```
Then PartialSinkConverter is a map:
```
SinkAC -> (SourceA, SinkBC)
SinkBC -> (SourceB, SinkAC)
```
We do not need to store a list of sources and partial sinks as values in the
above map, because it is undesirable and currently impossible to use the same
partial sink kinds in 2 multi-source rules. For example, it is impossible
to define another rule:
```
{
  SourceD, SourceE
  SinkC
}
```
because SinkC has been defined in the previous rule.

Reviewed By: arthaud

Differential Revision: D42422114

fbshipit-source-id: 273fa1de30b7e205bb6d98e1ddfb62f6d21850fc