Represent attributes and globals as targets in the model query execution logic
Summary: To implement `read_from_cache` and `WriteToCache`, we will use a cache of `Target.t`. To prevent having to translate `Target.t` back into the generic `target_or_reference` type, let's just use `Target.t` everywhere in the model query execution module. This simplies the API a bit. The only downside is that we now throw an exception in case it is called with a callable target when expecting a global or attribute.

Reviewed By: sinancepel

Differential Revision: D42444564

fbshipit-source-id: 7f6e81749331d816844394a14b9121aa9d28e31a