Allow call info "origin" to appear in "extra_traces" and Add field "message" into "extra_traces"
Summary:
Such that when an ExtraTraceSink is Origin, we still show subtraces for it.
In other words, for any Zoncolan UI frame, even if the first hop of its
subtrace is the same frame as itself, we still want to show the subtrace.

Field "message" allows showing the taint transforms when an ExtraTraceSink is
Origin -- when this happens, there is no subtrace to show and hence a message
would be helpful.

This diff replaces D42441833.

Reviewed By: arthaud

Differential Revision: D42457612

fbshipit-source-id: 1622dff274fb084e7fc26657e54870ade6f14e94