Rename statistics_logger to remote_logger
Summary:
The name of the module named statistics_logger suggests that it is for logging
statistics, e.g. the output of `pyre statistics` or some other kind of
aggregated data.

But in fact, this is just the remote logging logic used by the Pyre client when
telemetry is enabled (which it is not by default in open-source Pyre).

Renaming it to `remote_logger` makes it much more clear what this module is
actually all about, and makes it quite a bit easier to guess the module name.
In addition, it will prevent this module from coming up every time we search
for the code used by `pyre statistics`, which has been a minor source of
annoyance to me.

Reviewed By: sinancepel

Differential Revision: D42506361

fbshipit-source-id: b087bd4f10bc9b2321d155440d3afd8a66605e9a