Allow project identifier in pyre_configuration
Summary:
Background: Pyre and Pyls need a way to gate changes off of repo name. Pyre doesn't currently have a way to read repo id.

See [this doc](https://docs.google.com/document/d/1hCV5a8NHHz26ZOeBgY4gv_craUCE_-ZFMJkttAgV86s/edit#) for a further analysis. As of the update on the bottom, we decided we should go with option C of adding this name to the pyre_configuration.

This diff allows the field in the configuration - it does not add it yet since the client must be released before the warning will go away. It does not add it to the data structures on the client since as of now the VSCode extension is the only place we require this logic.

Other Options:
- We could do a simple upwards search of directory name until we hit a `.hg` or `.arcconfig` file like Pyls, but currently pyls would only know about `fbsource` and not `fbcode`.
- We could upwards search for a `.pyre_configuration` and use the name of that folder, however this could have conflicts and requires the configuration to not move.

Reviewed By: grievejia

Differential Revision: D42548626

fbshipit-source-id: f33aa3743f870443da02bcb10448f32dc8a0ad45