Rename "request handler" to "daemon querier"
Summary:
To me, a request handler is a dispatcher that accepts and responds
to requests. This does exist in Pyre language servers - it is the
method `PyreLanguageServer.handle_request`.

But the class formerly known as RequestHandler is not a request
handler... it doesn't deal with LSP requests at all but rather
queries to the daemon, and it isn't a handler in the sense of
a big dispatcher but rather an API with separate methods to call
when sending queries to the daemon.

The new name will hopefully make this clearer.

Reviewed By: grievejia

Differential Revision: D42617594

fbshipit-source-id: c2a15884fc9e37f44a59ba28697f659bc3c0b1a7