Support pytest's alias of `typing.overload`
Summary:
The latest release of pytest - probably the most widely used python testing library -
aliases `typing.overload` to `_pytest.compat.overload` and makes use of that decorator
on a number of key classes.

Pyre cannot handle this: we hardcode overload decorators in terms of fully-qualified
but *not* export-resolved names because we process overloads *before* the UGE is created,
and resolving exports is defined on the UGE.

The right solution would be to refactor and add an environment layer, which we intend
to do (see T142322930), but that will be a significant mini-project.

Suppressing type errors is not a good way to handle this, because by not recognizing
overloads in key testing-related code Pyre winds up complaining on many different
bits of *user* code, rather than the offending pytest code.

The short term fix is to just hardcode support.

Reviewed By: pradeep90

Differential Revision: D42644053

fbshipit-source-id: 13d560b68cced2af74e7cf9f7452517666ff7bbe