Add path to model query (2/n)
Summary:
We need to know the path for reporting errors of the right format for Model Query Sync.

This diff adds the path as an optional field to ModelQueries, and changed relevant code to support this change.

For model (not model query) verification errors, we do not add the path to the model. This is because for model verification errors that aren't model query errors, the errors can be validated during the parse step, while for model query errors, (no models generated, matching expected/unexpected models), we need to actually execute the model queries based on the analyzable callables.

Reviewed By: arthaud

Differential Revision: D42611691

fbshipit-source-id: 88fa574c4e8f82b6748d65bfaa08b4a6a380fbec