error when constructors attempt to shadow class-level variables with incompatible type
Summary: We had a soundness issue where instances of a class could shadow a class variable in construction that we'd ignore. See the test for an example - since attribute lookups prioritize the instance dict, if we view the class variable as the canonical declaration as we do today, we need to disallow differing annotations on the `__init__`.

Reviewed By: pradeep90

Differential Revision: D42641648

fbshipit-source-id: 5ed1053bfa3418be8f69e0cf0cc704e5a96588c9