Fix search-and-replace bugs
Summary:
I thought that Pyre check would make a refactor relatively safe;
there's always a chance of mistakes when doing a big search-and-replace
name change, but nearly every mistake is, in theory, likely to be caught
by static type checks.

Unfortunately, though, it turns out that Pyre gives up completely when
we assign a module to a variable - it treats the type as Any, which then
silently allows arbitrary bugs.

As a result, I got `querier` (the name of a variable) and `daemon_querier`
(the name of the module defining the class for that variable) mixed
up and never got any type errors.

We have no integration tests of language server code, only unit tests
that don't verify any of the production wiring logic, which is why
this escaped detection in unit tests.

Reviewed By: kinto0

Differential Revision: D42690154

fbshipit-source-id: 52ab8dc1cbec539eab2150cad5fcb82c91151ce3