Remove non-main issues in multi-source rules
Summary:
The changes are:
1. During the global fixpoint computation, whenever finding an issue where the
sink is a partial sink, we file an issue. By contrast, previously we file an
issue only if **both** issues for a multi-source rule are found.
2. After the global fixpoint computation, we traverse all issues related with
multi-source rules, in order to (1) remove the secondary issues in a
multi-source rule, (2) attach the secondary issues to the main issues as
subtraces (in the Zoncolan UI).

Reviewed By: arthaud

Differential Revision: D42513061

fbshipit-source-id: aead422e1097f1ca0c6afa03d51004b6b90cc1d1