Add a helper function to create language servers in tests
Summary:
The pyre server gets more complex when we support type errors for unsaved
changes, because now type errors can come from both the subscription (global
filesystem-based errors) and from queries (kicked off in response to didChange
requests).

This means that both the PyreLanguagesServer and the launch and subscribe
handler will need access to the type error handler. This, in turn, makes
*directly* constructing language servers in tests irritating, since we
need to pass the same type error handler to both the server and the background
subscription handler, and we want to use the same server state / output channel.

Making a helper function solves this problem - we can mostly preserve the same
interface and avoid repetitive code.

Reviewed By: grievejia

Differential Revision: D42691440

fbshipit-source-id: ee49357ccedb55ab5813896d502cb39e80a5f31d