Alpha-quality support for unsaved-changes type errors
Summary:
This commit adds basic support for type errors as-you-type in
the Pyre language server.

For users opted into the `pyre_enable_unsaved_changes` GK (at
the moment this is just me!), we will now include in didChange
event handling:
- an overlay update
- a query to get type errors from the overlay

The semantics of these type errors are a bit messy at the moment,
which is why I describe the support as "alpha quality" - it works but
for someone who doesn't understand the underlying data flows the
behavior might be surprising.

Whenever a watchman update happens, we will clear out the type
errors and re-populate based on what's on disk. This means that the
type errors in editor are aware of unsaved changes some of the time (after
a didChange event), but not at other times. This could be fixed
by tracking which files are dirty and sending type errors queries
inside the subscription-handling logic, which I will probably do
later on.

Since my biggest short-term goal is to use this in pyre_bento_server,
where watchman won't be needed at all, I'm deferring that for later.

I am also not worrying about debouncing for now. My current attitude
is that, if we're worried about long-tail performance (which we do
have evidence could be bad), we probably should just use an
`asyncio.Lock` to ensure that overlay updates don't queue up. But that
decision can be deferred for now; it might be worth trying to determine
whether queued up requests are ever a problem in Bento since that would
motivate fixing the worst-case behavior.

I'm still trying to decide whether to preemptively add logic for
clearing type errors *before* we send the overlay update. This
would be better UX in the case where the overlay update is
slow, because there wouldn't be stale errors sitting around.
But it also makes the errors "flicker" in and out of existence
which might be annoying. For now I'm leaving this out but
we could revisit the decision, it's pretty easy to add in.

Reviewed By: kinto0

Differential Revision: D42620985

fbshipit-source-id: 2774cf126ebd878a6e475b7e440dce1f45ad7491